+++ src/agent/tables.ts
@@ -1,14 +1,24 @@
 import {sortBy, pluck, all, identity} from "ramda"
 import {derived} from "svelte/store"
 import Cache from "src/util/cache"
+import {Tags} from "src/util/nostr"
 import {Table, listener, registry} from "src/agent/storage"
+import user from "src/agent/user"
 
-const sortByCreatedAt = sortBy(([k, x]) => -x.value.created_at)
+const sortByCreatedAt = sortBy(([k, x]) => x.value.created_at)
+const sortByLastSeen = sortBy(([k, x]) => x.value.last_seen)
 
-// Temporarily put no upper bound on people for 0.2.18 migration
 export const people = new Table("people", "pubkey", {
-  // cache: new Cache({max: 5000}),
-  // cache: new Cache({max: 20_000}),
+  cache: new Cache({
+    max: 10000,
+    // Don't delete the user's own profile or those of direct follows
+    sort: xs => {
+      const follows = Tags.wrap(user.getPetnames()).values().all()
+      const whitelist = new Set(follows.concat(user.getPubkey()))
+
+      return sortBy(([k, {lru, value}]) => (whitelist.has(value.pubkey) ? 0 : value.created_at), xs)
+    },
+  }),
 })
 
 export const userEvents = new Table("userEvents", "id", {
@@ -19,7 +29,9 @@ export const notifications = new Table("notifications", "id")
 export const contacts = new Table("contacts", "pubkey")
 export const rooms = new Table("rooms", "id")
 export const relays = new Table("relays", "url")
-export const routes = new Table("routes", "id")
+export const routes = new Table("routes", "id", {
+  cache: new Cache({max: 5000, sort: sortByLastSeen}),
+})
 
 listener.connect()