+++ src/app/listener.ts
@@ -102,7 +102,7 @@ const listen = async pubkey => {
   }
 
   listener = await network.listen({
-    delay: 10000,
+    delay: 5000,
     relays: getUserReadRelays(),
     filter: [
       {kinds: [1, 4], authors: [pubkey], since},