+++ src/partials/Anchor.svelte
@@ -15,9 +15,10 @@
     {"opacity-50": loading},
     switcher(type, {
       anchor: "underline",
-      button: "py-2 px-4 rounded bg-input text-accent whitespace-nowrap",
+      button:
+        "py-2 px-4 rounded bg-input text-accent whitespace-nowrap border border-solid border-gray-6 hover:bg-input-hover",
       "button-circle":
-        "w-10 h-10 flex justify-center items-center rounded-full bg-input text-accent whitespace-nowrap border border-solid border-gray-5",
+        "w-10 h-10 flex justify-center items-center rounded-full bg-input text-accent whitespace-nowrap border border-solid border-gray-6 hover:bg-input-hover",
       "button-accent":
         "py-2 px-4 rounded bg-accent text-white whitespace-nowrap border border-solid border-accent-light hover:bg-accent-light",
     })