+++ src/views/login/ConnectUser.svelte
@@ -67,7 +67,7 @@
 
             navigate("/notes/follows")
           } else {
-            pool.getConnection(relay.url).disconnect()
+            pool.disconnect(relay.url)
           }
         })
     }
@@ -115,7 +115,9 @@
       }}>here</Anchor
     >.
   </p>
-  {#if Object.values(currentRelays).length > 0}
+  {#if pool.forceRelays.length > 0}
+    <Spinner />
+  {:else if Object.values(currentRelays).length > 0}
     <p>Currently searching:</p>
     {#each Object.values(currentRelays) as relay}
       <div class="h-12">