+++ src/views/person/Notes.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+  import {timedelta} from "src/util/misc"
   import Feed from "src/views/feed/Feed.svelte"
   import {sampleRelays, getPubkeyWriteRelays} from "src/agent/relays"
 
@@ -8,4 +9,4 @@
   const filter = {kinds: [1], authors: [pubkey]}
 </script>
 
-<Feed {relays} {filter} parentsTimeout={3000} />
+<Feed {relays} {filter} parentsTimeout={3000} delta={timedelta(1, "days")} />