+++ vite.config.js
@@ -1,10 +1,10 @@
-import * as path from 'path'
-import {defineConfig} from 'vite'
-import {VitePWA} from 'vite-plugin-pwa'
-import mkcert from 'vite-plugin-mkcert'
-import sveltePreprocess from 'svelte-preprocess'
-import {svelte} from '@sveltejs/vite-plugin-svelte'
-import {nodePolyfills} from 'vite-plugin-node-polyfills'
+import * as path from "path"
+import {defineConfig} from "vite"
+import {VitePWA} from "vite-plugin-pwa"
+import mkcert from "vite-plugin-mkcert"
+import sveltePreprocess from "svelte-preprocess"
+import {svelte} from "@sveltejs/vite-plugin-svelte"
+import {nodePolyfills} from "vite-plugin-node-polyfills"
 
 export default defineConfig({
   server: {
@@ -15,8 +15,8 @@ export default defineConfig({
   },
   resolve: {
     alias: {
-      src: path.resolve(__dirname, 'src'),
-    }
+      src: path.resolve(__dirname, "src"),
+    },
   },
   plugins: [
     mkcert(),
@@ -24,13 +24,13 @@ export default defineConfig({
       protocolImports: true,
     }),
     VitePWA({
-      registerType: 'autoUpdate',
-      injectRegister: 'auto',
+      registerType: "autoUpdate",
+      injectRegister: "auto",
       manifest: {
-        name: 'Coracle',
-        short_name: 'Coracle',
-        description: 'Nostr, your way.',
-        theme_color: '#EB5E28',
+        name: "Coracle",
+        short_name: "Coracle",
+        description: "Nostr, your way.",
+        theme_color: "#EB5E28",
         icons: [
           {type: "image/png", sizes: "192x192", src: "/images/favicon/android-icon-192x192.png"},
           {type: "image/png", sizes: "512x512", src: "/images/favicon/android-icon-512x512.png"},
@@ -40,8 +40,7 @@ export default defineConfig({
     svelte({
       preprocess: sveltePreprocess(),
       onwarn: (warning, handler) => {
-
-        if (warning.code.startsWith('a11y-')) return
+        if (warning.code.startsWith("a11y-")) return
         if (warning.filename.includes("node_modules")) return
 
         handler(warning)